:recycle: Break user edit form into its own component.
