+++ web/src/use/use-user.ts
@@ -1,9 +1,9 @@
 import { type Ref, reactive, unref, watch, toRefs } from "vue"
 import { isNil } from "lodash"
 
-import usersApi, { type User } from "@/api/users-api"
+import usersApi, { type User, type UserUpdate } from "@/api/users-api"
 
-export { type User }
+export { type User, type UserUpdate }
 
 export function useUser(
   id: Ref<number | null | undefined>,
@@ -44,6 +44,32 @@ export function useUser(
     }
   }
 
+  async function save(customAttributes: UserUpdate): Promise<User> {
+    const staticId = unref(id)
+    if (isNil(staticId)) {
+      throw new Error("id is required")
+    }
+
+    const attributes = {
+      ...state.user,
+      ...customAttributes,
+    }
+
+    state.isLoading = true
+    try {
+      const { user } = await usersApi.update(staticId, attributes)
+      state.isErrored = false
+      state.user = user
+      return user
+    } catch (error) {
+      console.error("Failed to save current user:", error)
+      state.isErrored = true
+      throw error
+    } finally {
+      state.isLoading = false
+    }
+  }
+
   watch(
     () => unref(id),
     async (newId) => {
@@ -58,6 +84,7 @@ export function useUser(
     ...toRefs(state),
     fetch,
     refresh: fetch,
+    save,
   }
 }
